Drop unrelated drift from the protovalidate integration
This branch had accumulated changes unrelated to protovalidate from its old
base: a generated grpc-krpc yarn.lock, proto-google-common-protos API dumps, a
gradle-wrapper downgrade, reverts of edition-2023 packing in FieldParser and of
multi-type dispatch in the conformance driver, a conformance proto-export
restructure, and stale version reversions in libs.versions.toml.

Reset all of those to origin/main so the diff is just the protovalidate
integration: ProtoktMessageReflector, ProtoktStructValue, the Validator and
RuntimeContext wiring, the protokt-protovalidate friend-paths build config,
mavenLocal for the protovalidate-java snapshot, and the cel/protovalidate
version pins.

Conformance holds at 2840/2840.
