+++ settings.gradle.kts
@@ -77,7 +77,6 @@ include(
     "testing:protokt-generation-2",
     "testing:protobuf-java",
     "testing:protovalidate-conformance",
-    "testing:protovalidate-conformance:protos",
     "testing:protobufjs",
     "testing:testing-util",
