+++ testing/protovalidate-conformance/build.gradle.kts
@@ -13,6 +13,8 @@
  * limitations under the License.
  */
 
+import com.google.protobuf.gradle.GenerateProtoTask
+import com.google.protobuf.gradle.proto
 import org.gradle.api.distribution.plugins.DistributionPlugin.TASK_INSTALL_NAME
 import protokt.v1.gradle.protokt
 
@@ -27,11 +29,9 @@ protokt {
     formatOutput = false
 }
 
-
 dependencies {
     implementation(project(":protokt-protovalidate"))
     implementation(project(":protokt-reflect"))
-    implementation(project(":testing:protovalidate-conformance:protos"))
     implementation(kotlin("reflect"))
     implementation(libs.cel)
     implementation(libs.classgraph)
@@ -41,9 +41,38 @@ dependencies {
     testImplementation(libs.truth)
 }
 
+sourceSets.main {
+    proto {
+        srcDir(project.layout.buildDirectory.file("protovalidate/export"))
+    }
+}
+
 val protovalidateVersion = libs.versions.protovalidate.get()
 val gobin = project.layout.buildDirectory.file("gobin").get().asFile.absolutePath
+val bufExecutable = project.layout.buildDirectory.file("gobin/buf").get().asFile
 val conformanceExecutable = project.layout.buildDirectory.file("gobin/protovalidate-conformance").get().asFile
+
+val installBuf =
+    tasks.register<Exec>("installBuf") {
+        environment("GOBIN", gobin)
+        outputs.file(bufExecutable)
+        commandLine("go", "install", "github.com/bufbuild/buf/cmd/buf@v${libs.versions.buf.get()}")
+    }
+
+val downloadConformanceProtos =
+    tasks.register<Exec>("downloadConformanceProtos") {
+        dependsOn(installBuf)
+        commandLine(
+            bufExecutable,
+            "export",
+            "buf.build/bufbuild/protovalidate-testing:v$protovalidateVersion",
+            "--output=build/protovalidate/export"
+        )
+    }
+
+tasks.withType<GenerateProtoTask> {
+    dependsOn(downloadConformanceProtos)
+}
 
 val installConformance =
     tasks.register<Exec>("installProtovalidateConformance") {
